+++ app/gui/tsconfig.app.json
@@ -37,6 +37,7 @@
     "./src/components/ProtectedLayout.vue",
     "./src/components/RegistrationPage.vue",
     "./src/components/WithCurrentProject.vue",
+    "./src/components/projectLogData.ts",
     "./src/composables/appTitle.ts",
     "./src/composables/paywall/FeaturesConfiguration.ts",
     "./src/composables/paywall/index.ts",
